refactor: [relax] todo hook in useTodoManager
